Add project column to smoke_tests.
Update the smoke_tests table so that the primary project is
stored.

Add a migration that populates the initial value of the project column
in the smoke_tests table based on the the package_builder which uses
a non-default (non-master) branch.

Update the SmokeTest module so that it populates the project column
based on the PackageBuilder which uses a non-default (non-master)
branch.

Update the smoke_tests/index view to display the new project:branch
information instead of hard coding it to 'Nova'.